Procurementportal/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

requests
pandas
python-dotenv
beautifulsoup4
schedule
pymongo
openai
unstructured[all-docs]
python-docx
Pillow
sentence-transformers
lxml
docx
tqdm 
azure-functions
urllib3